Split Locator mapping
